% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliment2-package.R
\docType{package}
\name{compliment2-package}
\alias{compliment2}
\alias{compliment2-package}
\title{compliment2: Generate Random Compliments}
\description{
Generate randomized compliments, including compliments about coding in R.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
