% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliment-r.R
\name{compliment_r}
\alias{compliment_r}
\title{Generate a random compliment about coding in R}
\usage{
compliment_r(n = 1)
}
\arguments{
\item{n}{Number of compliments to generate.}
}
\value{
A character vector of length \code{n}.
}
\description{
Generate a random compliment about coding in R
}
\examples{
compliment_r()
}
