% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farewell.R
\name{farewell}
\alias{farewell}
\title{Generate a random goodbye}
\usage{
farewell(n = 1)
}
\arguments{
\item{n}{Number of goodbyes to generate.}
}
\value{
A character vector of length \code{n}.
}
\description{
Generate a random goodbye
}
\examples{
farewell()
}
